Auto logout user if token is invalid
