+++ src/helpers/mockStore.js
@@ -1,6 +1,7 @@
 import configureStore from "redux-mock-store";
+import thunk from "redux-thunk";
 
-const middlewares = [];
+const middlewares = [thunk];
 const mockStore = configureStore(middlewares);
 
 const testUser = {
