+++ src/store/logout.js
@@ -1,0 +1,3 @@
+import { createAction } from "@reduxjs/toolkit";
+
+export const logoutUser = createAction("LOGOUT_USER");
